Create Slicing DataFrames
Slicing rows
The Pennsylvania US election results data set that you have been using so far is ordered by county name. This means that county names can be sliced alphabetically. In this exercise, you're going to perform slicing on the county names of the election DataFrame from the previous exercises, which has been pre-loaded for you.

--------------------------------------------

Slicing columns
Similar to row slicing, columns can be sliced by value. 
In this exercise, your job is to slice column names from the Pennsylvania election results DataFrame using .loc[].

It has been pre-loaded for you as election, with the index set to 'county'.

----------------------------------------------


Subselecting DataFrames with lists
You can use lists to select specific row and column labels with the .loc[] accessor. 
In this exercise, your job is to select the counties ['Philadelphia', 'Centre', 'Fulton'] 
and the columns ['winner','Obama','Romney'] from the election DataFrame, 
which has been pre-loaded for you with the index set to 'county'.

-------------------------------------------------------------